feat(lyrics): sync stored options on first use
refactor(lyrics): move initLyrics & processLyrics to init.ts
fix(auto-scroll): stop jump on new song
